packagegroup-core-boot: add bootloader to EFI systems
Since the bootloader is now installed instead of added at image creation
time, this is necessary for booting.

Also set the default to grub-efi. This is done because a default of
grub-efi is already used by live-vm-common.bbclass, and in the event
that EFI_PROVIDER isn't defined errors would occur.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>