QemuRunner: Correct login message conditional statement
After a previous refactor of the QemuRunner class, the else clause
on the conditional statement to determine login success within the launch
method was moved one level below, causing user login messages, both for
success and failure, to be logged on a successful login (when
netconf==false), and nothing to be logged on failures.

This change corrects the else clause position.

Signed-off-by: Jair Gonzalez <jair.de.jesus.gonzalez.plascencia@linux.intel.com>