QemuRunner: Remove extra handler to avoid duplicated log lines
Each time an instance of QuemuRunner was launched from oe-selftest,
QemuRunner added a new StreamHandler for its logger. This caused an additional
line of duplicated messages on the output log each time a test using QEMU was
executed.

This change removes the addition of a handler within the QemuRunner class
so its parent (BitBake) catches and handles its events instead.

Signed-off-by: Jair Gonzalez <jair.de.jesus.gonzalez.plascencia@linux.intel.com>