fs-uuid.bbclass: rewrite get_fs_uuid function
I couldn't find anywhere it was being used, and in trying to use it it
just spit out an error.

It now generates a random UUID and puts it in the deploy dir for
consumption.

We also provide some utility variables in the bbclass for use in images
and bootloaders.

Note that we should add the rootfs-uuid file to file-checksums of
functions using the DISK_SIGNATURE_UUID variable, else it may become
incorrect due to caching.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>