devtool: refactor return for _extract_source()
Use a namedtuple to return information to the caller, since I've been
expanding that information we should avoid having to change all of the
calling code each time.

Additionally, it turned out that a bunch of the callers were checking
for None being returned in the initial_rev value, but that's no longer
possible, so tidy up the calling code.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>