kernel-devsrc: Use full mrproper to clean
By using the full mrproper clean target it also removes some generated files
that are not required on the target to build. It also helps to fix reprodudible
builds as the generated files could have different permissions causing the
check to fail.

Tested with building and booting kernel from source on the target

[YOCTO #12137]

Signed-off-by: Saul Wold <sgw@linux.intel.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>