lib/oe/path: implement is_path_under()
In a few places we have checks to see path A is under path B, by adding
/ to the end of the path B and then seeing if path A starts with the
suffixed path B. Unfortunately there are two potential flaws: (1) path A
needs to be suffixed with / as well or the directory itself won't match
(semantics perhaps, but in a lot of scenarios returning True is
correct); (2) you need to run os.path.abspath() on both paths first or
you will wrongly return False for some relative paths where you should
return True. Let's solve this once and for all by writing a function
that takes care of these and put it in oe.path.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>