sanity.bbclass: remove obsolete bblayers.conf updater code
The code in sanity.bbclass was executed for all distros, regardsless
whether they use the same numbering of bblayers.conf.sample as
OE-core.

The is problematic for distros which tracks their own LCONF_VERSION,
once they reach the 5 to 7 (inclusive) number range, the code from
sanity.bbclass kicks in and automatically updates the version to that
number without actually changing anything else in bblayers.conf, which
leaves the developer out-of-sync with bblayers.conf.sample.

We removed the automatic updater code blob and that will resolves the
issue of out-of-sync bblayers.conf. It will now notify users when
bblayers.conf is not valid and he needs to perform a manual update.

[YOCTO #11683]

Suggested-by: Patrick Ohly <patrick.ohly@intel.com>
Signed-off-by: Chang Rebecca Swee Fun <rebecca.swee.fun.chang@intel.com>