gettext-runtime: add recipe
This recipe is the "runtime" portion of gettext, which provides a libintl for C
runtimes without an integrated libintl (which glibc and musl do) and some tools
to use gettext from shell.

As this is rarely needed, we can simply split it into a separate recipe to speed
the build of gettext-tools.

Signed-off-by: Ross Burton <ross.burton@intel.com>