tinfoil.py: Check if BUILDDIR  or .devtoolbase exists
Check if either BUILDDIR or .devtoolbase exists in order to determine
if a tinfoil using script is running without having called oe-init-build-env
(or similar). If not, raise an exception that a tinfoil using application
can catch to produce an error message. Before, tinfoil would hang for 30s
before erroring out.

Fixes [YOCTO #12096]

Signed-off-by: Amanda Brindle <amanda.r.brindle@intel.com>