oepydevshell-internal.py: fix handling of unicode characters
In previous implementation, an AttributeError exception will be
raised when users attempted to change the pop-up terminal window
size (as reported).

The select call in the code was triggered and it mistakenly
think there is data to process in the pipe. We should only
attempt to decode if its data is not None.

Changed to use codecs module getreader function as it provides
better data stream decoding.

[YOCTO #11875]

Signed-off-by: Chang Rebecca Swee Fun <rebecca.swee.fun.chang@intel.com>