kernel-devsrc: minimize package size
By using the full mrproper clean target it also removes some generated files
that are not required on the target or can be regenerated on the target when
needed. In addition, remove all sources for all archs != ${ARCH}. This should
be harmless, as the sources are meant to allow user to reconfigure/rebuild the
existing kernel or build out-of-tree kernel modules. However, it will be no longer
possible to cross-compile the kernel for other architectures, but that was never
the real purpose of this package.

The changes also help to fix reprodudible builds as the generated files could have
different permissions causing the check to fail.

Tested with building and booting kernel from source on the target.
Tested with meta-skeleton/hello-mod
Tested with bitbake core-image-sato-sdk -c testimage
  (needs modified kernelmodule.py)
  - kernelmodule.KernelModuleTest.test_kernel_module - Testcase 1541: PASSED

[YOCTO #12137]

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>