image_types_wic.bbclass: add support for variable UUID
Random UUIDs must be known by both bootloaders and wic. This means that
it either needs to be set from a global space like bitbake.conf or
local.conf, or it must be written-to/retrieved-from a file, which is
then stored in a variable.

Somewhere in that translation, something is getting cached, causing the
dependant tasks to not get reparsed/reran. To prevent this, add the
file that stores the UUID to files-checksum, which corrects the task
when the UUID changes.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>