ovmf: refresh patch and use shell variable in recipe
In an eSDK environment, when we run 'devtool modify',
devtool will go through a series of tasks (fetch, unpack,
patch, etc). The bitbake variables within the recipe will
be replaced with a temporary directory path that was
created by 'devtool' for source extraction and patching.

This has caused an issue when we try to build the recipe
in eSDK environment, where the temporary file path has been
moved and deleted upon completion of do_patch task.

Now we used shell variable instead because we do not want a
temporary file path ended up in the source code at a early
stage of 'devtool modify'.

Refreshed patch 0002-ovmf-update-path-to-native-BaseTools.patch
to directly point the BaseTools path to a shell variable to
fix the building issue within eSDK.

[YOCTO #12153]

Signed-off-by: Chang Rebecca Swee Fun <rebecca.swee.fun.chang@intel.com>