systemd: change PACKAGECONFIG 'resolve' back to 'resolved'
When systemd was upgraded from 234 to 237, the PACKAGECONFIG item
'resolved' is changed to 'resolve', this is because meson_options.txt
uses the word 'resolve' instead of 'resolved'.

However, this causes trouble for users. Backward compatibility is obviously
more important, because we might have bbappend files in other layers
using this PACKAGECONFIG item.

So change the name back to 'resolved'.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>