runqueue.py: Initial implementation of per task process limits
On high core machines, in do_fetch, it is possible to DDoS your own machine.
So I added a method to limit any arbitrary task type to a certain number of
simultaneous threads.  (Similar to how BB_NUMBER_THREADS works in the general
case.)  The format is:

    do_fetch[number_threads] = "2"

This should be set globally.  If it is set in individual recipes it could
result in unpredictable behavior.

Note: a value for number_threads > BB_NUMBER_THREADS will have no effect.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>